remove unnecessary commands
