--- ncovstats/packages/ncovstats/cfr.js
@@ -1,72 +1,0 @@
-// jshint esversion: 9
-
-let cheerio;
-
-
-const coronaMeter = 'https://www.worldometers.info/coronavirus/';
-/**
- * @description null
- * @param {ParamsType} params list of command parameters
- * @param {?string} commandText text message
- * @param {!object} [secrets = {}] list of secrets
- * @return {Promise<SlackBodyType>} Response body
- */
-async function _command(params, commandText, secrets = {}) {
-
-
-  const axios = require('axios');
-
-  if (!cheerio) {
-    await install(['cheerio']);
-    cheerio = require('cheerio');
-  }
-
-  let response;
-  try {
-    response = await axios.get(coronaMeter);
-    if (response.status !== 200) {
-      throw err;
-    }
-  } catch (err) {
-    return null;
-  }
-
-  let result = {};
-  const html = cheerio.load(response.data);
-  const stat = html(`span.panel-title:contains(Fatality Rate)`);
-  if (stat.length) {
-    result = stat.parent().next().text();
-  }
-
-  return {
-    response_type: 'in_channel', // or `ephemeral` for private response
-    text: `FATALITY RATE (CFR) (very early estimate):\n ${result}`
-  };
-}
-
-
-// installs a set of npm packges
-async function install(pkgs) {
-  pkgs = pkgs.join(' ');
-  return new Promise((resolve, reject) => {
-    const { exec } = require('child_process');
-    exec(`npm install ${pkgs}`, (err, stdout, stderr) => {
-      if (err) reject(err);
-      else resolve();
-    });
-  });
-}
-
-/**
- * @typedef {object} SlackBodyType
- * @property {string} text
- * @property {'in_channel'|'ephemeral'} [response_type]
- */
-
-const main = async (args) => ({
-  body: await _command(args.params, args.commandText, args.__secrets || {}).catch(error => ({
-    response_type: 'ephemeral',
-    text: `Error: ${error.message}`
-  }))
-});
-module.exports = main;
